Refactor make and solve
Moved make and solve into their own modules. This will hopefully make it
easier to address changes to those specific units in the future.
