+++ python/src/utils.py
@@ -1,0 +1,5 @@
+
+import sys
+
+def eprint(*args, **kwargs):
+    print(*args, file=sys.stderr, **kwargs)
